Improve type inference during custom auto derivation
