+++ examples/src/main/scala/zio/config/examples/commandline/CommandLineComplex.scala
@@ -2,7 +2,7 @@ package zio.config.examples.commandline
 
 import zio.config._, ConfigDescriptor._
 
-object ComplexExample extends App {
+object CommandLineComplex extends App {
   val argss =
     "--conf -database.username=1 --conf -database.password=hi --conf.database.url=jdbc://xyz --conf -num_execs=10 --vault.username=3 --vault.password=10 --vault.something=11 --users 1 --users 2 --region 111,112"
