+++ examples/src/main/scala/zio/config/examples/magnolia/AutoDerivationSealedTrait.scala
