+++ examples/src/main/scala/zio/config/examples/magnolia/AutoDerivationSimple.scala
